requirements.txt
paho-mqtt==2.1.0
requests==2.28.1
python-dotenv==1.1.0